ml/requirements.txt
torch>=2.9,<2.10
torchvision>=0.24,<0.25
numpy>=2.1,<3
Pillow>=11.2,<13
